Refact: home은 v2로, newHome은 home으로
